feat:完善系统权限操作，保护默认角色和超级管理员权限不可侵犯

+++ src/main/java/com/mida/chromeext/modules/controller/manage/RolesController.java
@@ -1,15 +1,19 @@
 package com.mida.chromeext.modules.controller.manage;
 
 import com.mida.chromeext.components.shiro.PermisConstant;
+import com.mida.chromeext.components.shiro.RoleConstant;
 import com.mida.chromeext.modules.pojo.Role;
 import com.mida.chromeext.modules.service.RoleService;
 import com.mida.chromeext.modules.vo.ListQueryVo;
 import com.mida.chromeext.utils.Result;
+import com.mida.chromeext.utils.ResultCode;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -67,13 +71,24 @@ public class RolesController {
         return Result.ok(roleService.updateRole(role));
     }
 
+    /**
+     * 删除系统角色
+     * 系统默认的超级管理员和管理员角色禁止删除
+     * @param roleIds
+     * @return
+     */
     @DeleteMapping()
     @ApiOperation("删除角色列表")
     @RequiresPermissions(PermisConstant.DELETE_SYS_ROLE)
     public Result<Boolean> deleteList(@ApiParam("角色id列表") @RequestBody List<Integer> roleIds) {
+        List<Role>  roles = roleService.getRolesByIds(roleIds);
+        for (Role role : roles) {
+            if (role.getName().equals(RoleConstant.SUPER_ROLE) || role.getName().equals(RoleConstant.ADMIN_ROLE)) {
+                return Result.error(ResultCode.FORBIDDEN.code(), "Illegal Operation");
+            }
+        }
         try {
-            roleService.deleteRoles(roleIds);
-            return Result.ok();
+            return roleService.deleteRoles(roleIds) ? Result.ok(true) : Result.error();
         } catch (Exception e) {
             log.error(e.getMessage());
             return Result.error(e.getMessage());
@@ -85,8 +100,7 @@ public class RolesController {
     @RequiresPermissions(PermisConstant.ADD_PERMIS_TO_ROLE)
     public Result<Boolean> bindPermissionsToRoles(@PathVariable Integer roleId, @RequestBody List<Integer> permissionIds) {
         try {
-            roleService.addPermissionsToRole(roleId, permissionIds);
-            return Result.ok(true);
+            return roleService.addPermissionsToRole(roleId, permissionIds) > 0 ? Result.ok(true) : Result.error();
         } catch (Exception e) {
             log.error(e.getMessage());
             return Result.error(e.getMessage());
@@ -97,6 +111,16 @@ public class RolesController {
     @ApiOperation("通过角色rid和权限pid列表接触权限绑定")
     @RequiresPermissions(PermisConstant.REMOVE_PERMIS_OFF_ROLE)
     public Result<Boolean> unbindPermissionsOffRoles(@PathVariable Integer roleId, @RequestBody List<Integer> permissionIds) {
+        Role role = roleService.getRoleById(roleId);
+        // 禁止删除超级管理员权限
+        if (role.getName().equals(RoleConstant.SUPER_ROLE)) {
+            return Result.error(ResultCode.FORBIDDEN.code(), "Illegal Operation");
+        }
+        // 禁止非超级管理员删除管理员权限
+        Subject subject = SecurityUtils.getSubject();
+        if (role.getName().equals(RoleConstant.ADMIN_ROLE) && !subject.hasRole(RoleConstant.SUPER_ROLE)) {
+            return Result.error(ResultCode.FORBIDDEN.code(), "Illegal Operation");
+        }
         try {
             roleService.removePermissionsOfRole(roleId, permissionIds);
             return Result.ok(true);
